+++ back-end/src/main/java/com/tdd/backend/auth/encrypt/EncryptHelper.java
@@ -1,4 +1,4 @@
-package com.tdd.backend.auth.util;
+package com.tdd.backend.auth.encrypt;
 
 public interface EncryptHelper {
 	String encrypt(String password);
