+++ back-end/src/main/java/com/tdd/backend/auth/jwt/JwtTokenProvider.java
@@ -1,7 +1,7 @@
-package com.tdd.backend.auth.util;
+package com.tdd.backend.auth.jwt;
 
-import static com.tdd.backend.auth.util.JwtTokenProvider.JwtTokenRole.*;
-import static com.tdd.backend.auth.util.JwtTokenProvider.JwtTokenStatus.*;
+import static com.tdd.backend.auth.jwt.JwtTokenProvider.JwtTokenRole.*;
+import static com.tdd.backend.auth.jwt.JwtTokenProvider.JwtTokenStatus.*;
 import static org.apache.tomcat.util.codec.binary.Base64.*;
 
 import java.util.Base64;
