+++ back-end/src/main/java/com/tdd/backend/config/WebMvcConfig.java
@@ -7,7 +7,7 @@ import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 import com.tdd.backend.auth.AuthResolver;
-import com.tdd.backend.auth.util.JwtTokenProvider;
+import com.tdd.backend.auth.jwt.JwtTokenProvider;
 
 import lombok.RequiredArgsConstructor;
