+++ back-end/src/test/java/com/tdd/backend/user/UserServiceTest.java
@@ -10,8 +10,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.tdd.backend.auth.RefreshTokenStorage;
-import com.tdd.backend.auth.util.EncryptHelper;
+import com.tdd.backend.auth.jwt.RefreshTokenStorage;
+import com.tdd.backend.auth.encrypt.EncryptHelper;
 import com.tdd.backend.user.data.User;
 import com.tdd.backend.user.data.UserCreate;
 import com.tdd.backend.user.data.UserLogin;
